Fix export unit test pulling in Prisma client (CI module-not-found)
app/collab/__tests__/export.test.ts only exercises the pure renderers, but it
imported them from export.ts, which imports ~/lib/db → the generated Prisma
client. CI hasn't generated that client when unit tests run, so the suite failed
with ERR_MODULE_NOT_FOUND on ../generated/prisma/client.js.

Split the pure ProseMirror→HTML rendering into export-html.ts (no DB import).
export.ts keeps the DB-coupled decode functions and re-exports the pure ones, so
the export route's import surface is unchanged. export-pdf.ts now takes its PMNode
type from export-html. The test imports the pure functions from export-html, so
its module graph never reaches the DB client.

Verified by running the test with no generated Prisma client present (the CI
condition): passes.
